Moved http decorator tests into decorators/test_http.py.
