+++ tests/decorators/test_http.py
@@ -1,0 +1,22 @@
+from django.http import HttpRequest, HttpResponse, HttpResponseNotAllowed
+from django.test import SimpleTestCase
+from django.views.decorators.http import require_safe
+
+
+class RequireSafeDecoratorTest(SimpleTestCase):
+    def test_require_safe_accepts_only_safe_methods(self):
+        def my_view(request):
+            return HttpResponse("OK")
+
+        my_safe_view = require_safe(my_view)
+        request = HttpRequest()
+        request.method = "GET"
+        self.assertIsInstance(my_safe_view(request), HttpResponse)
+        request.method = "HEAD"
+        self.assertIsInstance(my_safe_view(request), HttpResponse)
+        request.method = "POST"
+        self.assertIsInstance(my_safe_view(request), HttpResponseNotAllowed)
+        request.method = "PUT"
+        self.assertIsInstance(my_safe_view(request), HttpResponseNotAllowed)
+        request.method = "DELETE"
+        self.assertIsInstance(my_safe_view(request), HttpResponseNotAllowed)
